# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wraith API",
    "version": "1.0.0",
    "description": "REST API documentation for the Wraith token transfer indexer.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/accounts/{address}/transfers": {
      "get": {
        "summary": "Account transfers",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "example": "GABCDEFGHIJKLMNOPQRSTUVWXYZ"
            },
            "required": true,
            "name": "address",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Token contract ID to filter by"
            },
            "required": false,
            "description": "Token contract ID to filter by",
            "name": "contractId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^C[A-Z2-7]{55}$",
              "example": "CB64D3G7SM2RTH6ISYIG4P2IYYD6J2OFR6B"
            },
            "required": false,
            "name": "token",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true
            },
            "required": false,
            "name": "fromLedger",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true
            },
            "required": false,
            "name": "toLedger",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "format": "date-time",
              "description": "Inclusive lower bound on ledgerClosedAt"
            },
            "required": false,
            "description": "Inclusive lower bound on ledgerClosedAt",
            "name": "fromDate",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "format": "date-time",
              "description": "Inclusive upper bound on ledgerClosedAt"
            },
            "required": false,
            "description": "Inclusive upper bound on ledgerClosedAt",
            "name": "toDate",
            "in": "query"
          },
          {
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string"
              },
              "description": "Comma-separated list of event types"
            },
            "required": false,
            "description": "Comma-separated list of event types",
            "name": "eventType",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true,
              "default": 50,
              "description": "Page size"
            },
            "required": false,
            "description": "Page size",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true,
              "default": 0,
              "description": "Pagination offset"
            },
            "required": false,
            "description": "Pagination offset",
            "name": "offset",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Opaque pagination cursor"
            },
            "required": false,
            "description": "Opaque pagination cursor",
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "OData filter expression"
            },
            "required": false,
            "description": "OData filter expression",
            "name": "$filter",
            "in": "query"
          },
          {
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string"
              },
              "description": "Comma-separated field list"
            },
            "required": false,
            "description": "Comma-separated field list",
            "name": "$select",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "transfers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "contractId": {
                            "type": "string",
                            "pattern": "^C[A-Z2-7]{55}$",
                            "example": "CB64D3G7SM2RTH6ISYIG4P2IYYD6J2OFR6B"
                          },
                          "eventType": {
                            "type": "string",
                            "enum": [
                              "transfer",
                              "mint",
                              "burn",
                              "clawback"
                            ]
                          },
                          "fromAddress": {
                            "type": "string",
                            "nullable": true,
                            "minLength": 1,
                            "example": "GABCDEFGHIJKLMNOPQRSTUVWXYZ"
                          },
                          "toAddress": {
                            "type": "string",
                            "nullable": true,
                            "minLength": 1,
                            "example": "GABCDEFGHIJKLMNOPQRSTUVWXYZ"
                          },
                          "amount": {
                            "type": "string"
                          },
                          "displayAmount": {
                            "type": "string"
                          },
                          "ledger": {
                            "type": "integer"
                          },
                          "ledgerClosedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "txHash": {
                            "type": "string"
                          },
                          "eventId": {
                            "type": "string"
                          },
                          "direction": {
                            "type": "string",
                            "enum": [
                              "incoming",
                              "outgoing"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "contractId",
                          "eventType",
                          "fromAddress",
                          "toAddress",
                          "amount",
                          "displayAmount",
                          "ledger",
                          "ledgerClosedAt",
                          "txHash",
                          "eventId"
                        ]
                      }
                    }
                  },
                  "required": [
                    "total",
                    "limit",
                    "offset",
                    "transfers"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/nfts/transfers": {
      "get": {
        "summary": "NFT transfers",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "NFT contract ID to filter by"
            },
            "required": false,
            "description": "NFT contract ID to filter by",
            "name": "contract",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "NFT token identifier"
            },
            "required": false,
            "description": "NFT token identifier",
            "name": "token_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Filter by sender or recipient address"
            },
            "required": false,
            "description": "Filter by sender or recipient address",
            "name": "address",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true
            },
            "required": false,
            "name": "fromLedger",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true
            },
            "required": false,
            "name": "toLedger",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true,
              "default": 50,
              "description": "Page size"
            },
            "required": false,
            "description": "Page size",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true,
              "default": 0,
              "description": "Pagination offset"
            },
            "required": false,
            "description": "Pagination offset",
            "name": "offset",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Opaque pagination cursor"
            },
            "required": false,
            "description": "Opaque pagination cursor",
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "OData filter expression"
            },
            "required": false,
            "description": "OData filter expression",
            "name": "$filter",
            "in": "query"
          },
          {
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string"
              },
              "description": "Comma-separated field list"
            },
            "required": false,
            "description": "Comma-separated field list",
            "name": "$select",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "transfers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "contractId": {
                            "type": "string",
                            "pattern": "^C[A-Z2-7]{55}$",
                            "example": "CB64D3G7SM2RTH6ISYIG4P2IYYD6J2OFR6B"
                          },
                          "tokenId": {
                            "type": "string"
                          },
                          "fromAddress": {
                            "type": "string",
                            "nullable": true,
                            "minLength": 1,
                            "example": "GABCDEFGHIJKLMNOPQRSTUVWXYZ"
                          },
                          "toAddress": {
                            "type": "string",
                            "nullable": true,
                            "minLength": 1,
                            "example": "GABCDEFGHIJKLMNOPQRSTUVWXYZ"
                          },
                          "ledger": {
                            "type": "integer"
                          },
                          "ledgerClosedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "txHash": {
                            "type": "string"
                          },
                          "eventId": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "contractId",
                          "tokenId",
                          "fromAddress",
                          "toAddress",
                          "ledger",
                          "ledgerClosedAt",
                          "txHash",
                          "eventId",
                          "createdAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "total",
                    "limit",
                    "offset",
                    "transfers"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}